agent_forge/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM integrations
openai>=1.3.0
anthropic>=0.5.0
google-generativeai>=0.1.0
cohere>=4.0.0
requests>=2.25.0  # For Ollama API

# Database integrations
supabase>=1.0.0
chromadb>=0.4.0
pinecone-client>=2.2.0
langchain>=0.0.200

# Embedding models
sentence-transformers>=2.2.0

# UI
streamlit>=1.26.0

# Utils
loguru>=0.7.0
pyyaml>=6.0.0
